+++ docs/realm.js
@@ -133,6 +133,14 @@ class Realm {
     static automaticSyncConfiguration(user) {}
 
     /**
+     * Creates a template object where all optional fields are `undefined` and all required fields have the default
+     * value for the given data type, i.e. `0`, false and `""`.
+     *
+     * @param {Realm~ObjectSchema} schema object describing the class
+     */
+    static createTemplateObject(objectSchema) {}
+
+    /**
      * Closes this Realm so it may be re-opened with a newer schema version.
      * All objects and collections from this Realm are no longer valid after calling this method.
      */