+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+2.9.0 Release notes (YYYY-MM-DD)
+=============================================================
+### Compatibility
+* Sync protocol: 24
+* Server-side history format: 4
+* File format: 7
+* Realm Object Server: 3.0.0 or later
+
+### Enhancements
+* Added `Realm.createTemplateObject(objectSchema)' (#1870).
+
+
 2.8.4 Release notes (2018-6-15)
 =============================================================
 ### Compatibility